CSL_Item note property and append methods
Includes suggestions by from Evgeniy <e.pogrebnyak@gmail.com>
https://github.com/manubot/manubot/pull/157#issuecomment-545310383
